+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---------------
 - add save-button to individual heroes
 - add buttons to set hero level from experience and vice versa
+- add info texts on artifact bonuses to hero main attributes
 - fix game version name not being fully visible (issue #11)
 - fix not being able to modify experience points over 2^30
 - cap hero level at 75 (74 in Horn of the Abyss)